Google functionality

+++ app/src/main/java/com/example/memokeeper/MainActivity.java
@@ -1,18 +1,23 @@
 package com.example.memokeeper;
 
+import android.Manifest;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.database.Cursor;
 import android.os.AsyncTask;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 
 import com.example.memokeeper.Constants.REQUEST_CODE;
 import com.example.memokeeper.DatabaseHelper.MemoContract;
@@ -20,8 +25,11 @@ import com.example.memokeeper.MainScreen.MemoAdapter;
 import com.example.memokeeper.MainScreen.MemoInfo;
 import com.example.memokeeper.MainScreen.VerticalSpaceItemDecoration;
 import com.example.memokeeper.MemoEditor.MemoEditActivity;
+import com.example.memokeeper.ProfilePage.ProfilePageActivity;
+import com.example.memokeeper.ProfilePage.SignInActivity;
 import com.example.memokeeper.Utilities.DateUtils;
 import com.example.memokeeper.Utilities.PathUtils;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -33,6 +41,9 @@ public class MainActivity extends AppCompatActivity{
 
     final private Context context = this;
 
+    private GoogleSignInAccount user = null;
+
+    final private int PERMISSION_REQUEST = 1;
     private MenuInflater inflater;
     private MemoAdapter memoAdapter;
     private ArrayList<MemoInfo> memo;
@@ -40,14 +51,17 @@ public class MainActivity extends AppCompatActivity{
     private MemoAsync memoGrabber = new MemoAsync();
     private String currentUnusedHash;
     private boolean newMemoCreated = false;
+    private MemoContract.MemoDbHelper dbHelper;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        checkPermission();
 
         Toolbar mainScreenToolbar = findViewById(R.id.mainScreenToolbar);
+        dbHelper = new MemoContract().new MemoDbHelper(context);
         memoList = findViewById(R.id.memoList);
         memo = new ArrayList<>();
         memoGrabber.execute("");
@@ -65,11 +79,12 @@ public class MainActivity extends AppCompatActivity{
     @Override
     protected void onDestroy() {
         super.onDestroy();
+        dbHelper.close();
         if (newMemoCreated) {
             File deleteFolder = new File(getFilesDir().getAbsolutePath(), currentUnusedHash);
             deleteFolder.delete();
             newMemoCreated = false;
-            Log.d("Destroy", "Temporary folder " + currentUnusedHash + " deleted!");
+            Log.d("Destroy temp folder", "Temporary folder " + currentUnusedHash + " deleted!");
         }
 
     }
@@ -95,6 +110,18 @@ public class MainActivity extends AppCompatActivity{
                 startActivityForResult(INTENT, REQUEST_CODE.MEMO_EDIT);
                 return true;
 
+            case R.id.action_google_signin:
+                if (isSignedIn()) {
+                    Intent PROFILE = new Intent(context, ProfilePageActivity.class);
+                    PROFILE.putExtra("user", user);
+                    startActivityForResult(PROFILE, REQUEST_CODE.VIEW_PROFILE);
+                }
+                else {
+                    Intent PROFILE = new Intent(context, SignInActivity.class);
+                    startActivityForResult(PROFILE, REQUEST_CODE.SIGN_IN);
+                }
+                return true;
+
             default:
                 return false;
         }
@@ -114,21 +141,35 @@ public class MainActivity extends AppCompatActivity{
                 if (pos == -1) {
                     memo.add(newMemo);
                     memoAdapter.notifyItemInserted(memo.size() - 1);
-                    MemoContract.MemoDbHelper dbHelper = new MemoContract().new MemoDbHelper(context);
                     dbHelper.addNewMemo(newMemo);
                 } else {
                     memo.set(pos, newMemo);
                     memoAdapter.notifyItemChanged(pos);
-                    MemoContract.MemoDbHelper dbHelper = new MemoContract().new MemoDbHelper(context);
                     dbHelper.updateMemo(newMemo);
 
                 }
             }
-            else {
+            else if (ResultCode == RESULT_CANCELED) {
                 if (newMemoCreated) {
                     File deleteFolder = new File(getFilesDir().getAbsolutePath(), currentUnusedHash);
                     deleteFolder.delete();
                     newMemoCreated = false;
+                }
+            }
+        }
+        if (RequestCode == REQUEST_CODE.SIGN_IN) {
+            if (ResultCode == RESULT_OK) {
+                user = data.getParcelableExtra("account");
+                Log.d("Account", user.getEmail() + "");
+                Log.d("Account", user.getDisplayName() + "");
+                Log.d("Account", user.getGivenName() + "");
+                Log.d("Account", user.getId() + "");
+            }
+        }
+        if (RequestCode == REQUEST_CODE.VIEW_PROFILE) {
+            if (ResultCode == RESULT_OK) {
+                if (data.getBooleanExtra("Sign out", false)) {
+                    user = null;
                 }
             }
         }
@@ -143,8 +184,7 @@ public class MainActivity extends AppCompatActivity{
     private class MemoAsync extends AsyncTask<String, Void, ArrayList<MemoInfo>> {
         @Override
         protected ArrayList<MemoInfo> doInBackground(String... url) {
-            MemoContract.MemoDbHelper fetchMemo = new MemoContract().new MemoDbHelper(context);
-            Cursor allMemo = fetchMemo.getAllMemo();
+            Cursor allMemo = dbHelper.getAllMemo();
             ArrayList<MemoInfo> memoList = new ArrayList<>();
             if (allMemo.moveToFirst()) {
                 do {
@@ -156,11 +196,44 @@ public class MainActivity extends AppCompatActivity{
                     memoList.add(new MemoInfo(memoTitle, date, memoContent, memoAttachment, hash));
                 } while (allMemo.moveToNext());
             }
+            allMemo.close();
             return memoList;
         }
 
         protected void onPostExecute(ArrayList<MemoInfo> result) {
             updateView(result);
+        }
+    }
+
+    private boolean isSignedIn() {
+        return (user != null);
+    }
+
+    private void checkPermission() {
+        if (ContextCompat.checkSelfPermission(MainActivity.this,
+                Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
+            if (ActivityCompat.shouldShowRequestPermissionRationale(MainActivity.this, Manifest.permission.READ_EXTERNAL_STORAGE)) {
+                ActivityCompat.requestPermissions(MainActivity.this,
+                        new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, PERMISSION_REQUEST);
+
+            } else {
+                ActivityCompat.requestPermissions(MainActivity.this,
+                        new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, PERMISSION_REQUEST);
+
+            }
+        }
+
+        if (ContextCompat.checkSelfPermission(MainActivity.this,
+                Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
+            if (ActivityCompat.shouldShowRequestPermissionRationale(MainActivity.this, Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+                ActivityCompat.requestPermissions(MainActivity.this,
+                        new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, PERMISSION_REQUEST);
+
+            } else {
+                ActivityCompat.requestPermissions(MainActivity.this,
+                        new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, PERMISSION_REQUEST);
+
+            }
         }
     }
 }
